+++ src/packages/user/change-password/index.ts
@@ -1,0 +1,1 @@
+export * from './modal/index.js';
