+++ src/packages/user/change-password/modal/index.ts
@@ -1,0 +1,1 @@
+export * from './change-password-modal.token.js';
