+++ src/packages/user/change-password/modal/manifests.ts
@@ -1,0 +1,8 @@
+export const manifests: Array<UmbExtensionManifest> = [
+	{
+		type: 'modal',
+		alias: 'Umb.Modal.ChangePassword',
+		name: 'Change Password Modal',
+		js: () => import('./change-password-modal.element.js'),
+	},
+];
